+++ NEWS.md
@@ -58,6 +58,10 @@ __Version 0.10b2 - 2019-05-30__
 
 __Version 0.10b3 - ... __
 
+- CHANGE: entity query - exclude DXF types from `'*'` search, by appending type name with a preceding '!' e.g. query for 
+  all entities except LINE = `"* !LINE"`
+- CHANGE: entity query - removed regular expression support for type name match
+
 Version 0.9 - 2019-02-24
 ------------------------